+++ Packages/webxr-input-profiles-loader/CHANGELOG.md
@@ -5,11 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- InputProfileLoader - Can set profilesUrl only on LoadProfilesList.
+- Moved ShaderVariants folder to package Runtime folder from Samples folder.
+- Use custom DeferAgent for loading gLTF models, for better loading performance.
+- Make sure to download a profile only once.
+- Updated glTFast package to 2.5.0.
 
 ## [0.2.0] - 2020-12-06
 ### Added
-- Support profilesList.json file
+- Support profilesList.json file.
 
 ## [0.1.0] - 2020-11-21
 ### Added
-- This package
+- This package.